Fix lint in tests
