+++ tests/integration/network/test_publisher.py
@@ -14,7 +14,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 from async_timeout import timeout
-import pytest
 
 from cylc.flow.network.subscriber import (
     WorkflowSubscriber,
