+++ tests/unit/test_taskdef.py
@@ -15,15 +15,12 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from cylc.flow.config import WorkflowConfig
-from cylc.flow.taskdef import generate_graph_parents
-from cylc.flow.cycling.iso8601 import ISO8601Point
 from cylc.flow.cycling.integer import IntegerPoint
+from cylc.flow.cycling.iso8601 import ISO8601Point
+from cylc.flow.taskdef import generate_graph_parents
 
 
-from .test_config import tmp_flow_config   # noqa: F401
-
-
-def test_generate_graph_parents_1(tmp_flow_config):   # noqa: F811
+def test_generate_graph_parents_1(tmp_flow_config):
     """Test that parents are only generated from valid recurrences."""
     id_ = 'pan-galactic'
     flow_file = tmp_flow_config(
@@ -64,7 +61,7 @@ def test_generate_graph_parents_1(tmp_flow_config):   # noqa: F811
         ]
 
 
-def test_generate_graph_parents_2(tmp_flow_config):   # noqa: F811
+def test_generate_graph_parents_2(tmp_flow_config):
     """Test inferred parents are valid w.r.t to their own recurrences."""
     id_ = 'gargle-blaster'
     flow_file = tmp_flow_config(
